feat: Phase 4 - Improvements & Report Enhancements
- ReportService: resolve college/department/supervisor names instead of IDs in report title
- StudentEvaluationService: populate EvaluatedByName with actual user name via IUserRepository
- CommitteeService: add conflict-of-interest check (supervisor cannot join committee evaluating their team)
- Delete duplicate EvaluationsView/ViewModel and remove all references from navigation
- Fix ClampCompletion to accept decimal input, fix CreatedByUserId nullable handling

+++ MasarApp/Masar.Application/Services/ProjectService.cs
@@ -341,11 +341,11 @@ public class ProjectService : IProjectService
         return Result.Success();
     }
 
-    private static int ClampCompletion(int value)
+    private static int ClampCompletion(decimal value)
     {
         if (value < 0) return 0;
         if (value > 100) return 100;
-        return value;
+        return (int)value;
     }
 
     private async Task<Result> EnsureSupervisionCapacity(int supervisorId, int? excludeProjectId, CancellationToken cancellationToken)
